AIPhotoRecognitionTool/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "photofilter"
version = "1.0.0"
description = "AI-powered photo organization system with advanced deduplication features"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "PhotoFilter Development Team"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=1.9.0",
    "torchvision>=0.10.0",
    "ultralytics>=8.0.0",
    "Pillow>=8.0.0",
    "opencv-python>=4.5.0",
    "seaborn>=0.11.0",
    "matplotlib>=3.3.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "tqdm>=4.62.0",
    "imagehash>=4.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.8",
]

[project.scripts]
photofilter = "photofilter.gui.main_window:main"

[project.urls]
Homepage = "https://github.com/rulfe/AIPhotoRecognitionTool"
Repository = "https://github.com/rulfe/AIPhotoRecognitionTool"
Issues = "https://github.com/rulfe/AIPhotoRecognitionTool/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
photofilter = ["config/*.json"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
]

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [".git", "__pycache__", "build", "dist"]
